Agrego mi segunda version del código Proyecto.py
